+++ src/test/java/io/supertokens/test/oauth/api/OAuthClientsAPITest.java
@@ -22,12 +22,16 @@ import com.google.gson.JsonPrimitive;
 import io.supertokens.ProcessState;
 import io.supertokens.httpRequest.HttpRequest;
 import io.supertokens.pluginInterface.RECIPE_ID;
+import io.supertokens.pluginInterface.exceptions.StorageQueryException;
+import io.supertokens.pluginInterface.multitenancy.AppIdentifier;
+import io.supertokens.pluginInterface.oauth.exceptions.OAuth2ClientAlreadyExistsForAppException;
 import io.supertokens.pluginInterface.oauth.sqlStorage.OAuthSQLStorage;
 import io.supertokens.storageLayer.StorageLayer;
 import io.supertokens.test.TestingProcessManager;
 import io.supertokens.test.Utils;
 import io.supertokens.test.httpRequest.HttpRequestForTesting;
 import io.supertokens.test.httpRequest.HttpResponseException;
+import jakarta.servlet.ServletException;
 import org.junit.*;
 import org.junit.rules.TestRule;
 
@@ -224,4 +228,245 @@ public class OAuthClientsAPITest {
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
     }
+
+    @Test
+    public void testClientUpdatePatch()
+            throws StorageQueryException, IOException, OAuth2ClientAlreadyExistsForAppException,
+            io.supertokens.test.httpRequest.HttpResponseException, InterruptedException {
+
+        String[] args = {"../"};
+
+        this.process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        String clientId = "6030f07e-c8ef-4289-80c9-c18e0bf4f679";
+        String propToChangeKey = "clientName";
+        String newValue = "Jozef";
+
+        OAuthSQLStorage oAuthStorage = (OAuthSQLStorage) StorageLayer.getStorage(process.getProcess());
+
+        AppIdentifier testApp = new AppIdentifier("", "");
+        oAuthStorage.addClientForApp(testApp, clientId);
+
+        JsonObject requestBody = new JsonObject();
+        requestBody.addProperty("clientId", clientId);
+        requestBody.addProperty(propToChangeKey, newValue);
+
+        JsonObject actualResponse = HttpRequestForTesting.sendJsonPATCHRequest(process.getProcess(),
+                "http://localhost:3567/recipe/oauth/clients", requestBody);
+
+        assertEquals("OK", actualResponse.get("status").getAsString());
+        assertTrue(actualResponse.has("client"));
+
+        JsonObject updatedClient = actualResponse.get("client").getAsJsonObject();
+
+        assertTrue(updatedClient.has(propToChangeKey));
+        assertEquals(newValue, updatedClient.get(propToChangeKey).getAsString());
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testClientUpdatePatch_multipleFields()
+            throws StorageQueryException, IOException, OAuth2ClientAlreadyExistsForAppException,
+            io.supertokens.test.httpRequest.HttpResponseException, InterruptedException {
+
+        String[] args = {"../"};
+
+        this.process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        String clientId = "6030f07e-c8ef-4289-80c9-c18e0bf4f679";
+        String propToChangeKey = "clientName";
+        String newValue = "Jozef2";
+
+        String listPropToChange = "grantTypes";
+        JsonArray newListValue = new JsonArray();
+        newListValue.add(new JsonPrimitive("test1"));
+        newListValue.add(new JsonPrimitive("test2"));
+
+        OAuthSQLStorage oAuthStorage = (OAuthSQLStorage) StorageLayer.getStorage(process.getProcess());
+
+        AppIdentifier testApp = new AppIdentifier("", "");
+        oAuthStorage.addClientForApp(testApp, clientId);
+
+        JsonObject requestBody = new JsonObject();
+        requestBody.addProperty("clientId", clientId);
+        requestBody.addProperty(propToChangeKey, newValue);
+        requestBody.add(listPropToChange, newListValue);
+
+        JsonObject actualResponse = HttpRequestForTesting.sendJsonPATCHRequest(process.getProcess(),
+                "http://localhost:3567/recipe/oauth/clients", requestBody);
+
+        assertEquals("OK", actualResponse.get("status").getAsString());
+        assertTrue(actualResponse.has("client"));
+
+        JsonObject updatedClient = actualResponse.get("client").getAsJsonObject();
+
+        assertTrue(updatedClient.has(propToChangeKey));
+        assertEquals(newValue, updatedClient.get(propToChangeKey).getAsString());
+
+        assertTrue(updatedClient.has(listPropToChange));
+        assertEquals(newListValue.getAsJsonArray(), updatedClient.get(listPropToChange).getAsJsonArray());
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testClientUpdatePatch_missingClientIdResultsInError()
+            throws StorageQueryException, IOException, OAuth2ClientAlreadyExistsForAppException,
+            io.supertokens.test.httpRequest.HttpResponseException, InterruptedException {
+
+        String[] args = {"../"};
+
+        this.process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        String clientId = "6030f07e-c8ef-4289-80c9-c18e0bf4f679";
+        String propToChangeKey = "clientName";
+        String newValue = "Jozef2";
+
+        String listPropToChange = "grantTypes";
+        JsonArray newListValue = new JsonArray();
+        newListValue.add(new JsonPrimitive("test1"));
+        newListValue.add(new JsonPrimitive("test2"));
+
+        OAuthSQLStorage oAuthStorage = (OAuthSQLStorage) StorageLayer.getStorage(process.getProcess());
+
+        AppIdentifier testApp = new AppIdentifier("", "");
+        oAuthStorage.addClientForApp(testApp, clientId);
+
+        JsonObject requestBody = new JsonObject();
+        //note the missing client Id
+        requestBody.addProperty(propToChangeKey, newValue);
+        requestBody.add(listPropToChange, newListValue);
+
+        io.supertokens.test.httpRequest.HttpResponseException expected = assertThrows(io.supertokens.test.httpRequest.HttpResponseException.class,
+                () -> {        HttpRequestForTesting.sendJsonPATCHRequest(process.getProcess(),
+                "http://localhost:3567/recipe/oauth/clients", requestBody);
+        });
+
+        assertEquals(400, expected.statusCode);
+        assertEquals("Http error. Status Code: 400. Message: Field name `clientId` is missing in JSON input\n", expected.getMessage());
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testClientUpdatePatch_hydraErrortResultsInError()
+            throws StorageQueryException, IOException, OAuth2ClientAlreadyExistsForAppException,
+            io.supertokens.test.httpRequest.HttpResponseException, InterruptedException {
+
+        String[] args = {"../"};
+
+        this.process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        String clientId = "6030f07e-c8ef-4289-80c9-c18e0bf4f679";
+        String propToChangeKey = "clientName";
+        String newValue = "Jozef2";
+
+        String notAlistPropToChange = "scope";
+        JsonArray newListValue = new JsonArray();
+        newListValue.add(new JsonPrimitive("test1"));
+        newListValue.add(new JsonPrimitive("test2"));
+
+        OAuthSQLStorage oAuthStorage = (OAuthSQLStorage) StorageLayer.getStorage(process.getProcess());
+
+        AppIdentifier testApp = new AppIdentifier("", "");
+        oAuthStorage.addClientForApp(testApp, clientId);
+
+        JsonObject requestBody = new JsonObject();
+        requestBody.addProperty("clientId", clientId);
+        requestBody.addProperty(propToChangeKey, newValue);
+        requestBody.add(notAlistPropToChange, newListValue);
+
+        HttpResponseException expected = assertThrows(HttpResponseException.class, () -> {
+            HttpRequestForTesting.sendJsonPATCHRequest(process.getProcess(),
+                            "http://localhost:3567/recipe/oauth/clients", requestBody);
+        });
+
+        assertEquals("Http error. Status Code: 500. Message: Internal Error\n", expected.getMessage());
+        assertEquals(500, expected.statusCode);
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testClientUpdatePatch_invalidInputDataResultsInError()
+            throws StorageQueryException, IOException, OAuth2ClientAlreadyExistsForAppException,
+            io.supertokens.test.httpRequest.HttpResponseException, InterruptedException {
+
+        String[] args = {"../"};
+
+        this.process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        String clientId = "6030f07e-c8ef-4289-80c9-c18e0bf4f679";
+        String propToChangeKey = "clientName";
+        String newValue = "Jozef2";
+
+        String notAlistPropToChange = "allowed_cors_origins";
+        JsonArray newListValue = new JsonArray();
+        newListValue.add(new JsonPrimitive("*"));
+        newListValue.add(new JsonPrimitive("appleTree"));
+
+        OAuthSQLStorage oAuthStorage = (OAuthSQLStorage) StorageLayer.getStorage(process.getProcess());
+
+        AppIdentifier testApp = new AppIdentifier("", "");
+        oAuthStorage.addClientForApp(testApp, clientId);
+
+        JsonObject requestBody = new JsonObject();
+        requestBody.addProperty("clientId", clientId);
+        requestBody.addProperty(propToChangeKey, newValue);
+        requestBody.add(notAlistPropToChange, newListValue);
+
+        io.supertokens.test.httpRequest.HttpResponseException expected = assertThrows(io.supertokens.test.httpRequest.HttpResponseException.class,
+                () -> {        HttpRequestForTesting.sendJsonPATCHRequest(process.getProcess(),
+                        "http://localhost:3567/recipe/oauth/clients", requestBody);
+                });
+
+        assertEquals(400, expected.statusCode);
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testClientUpdatePatch_notExistingClientResultsInNotFound()
+            throws StorageQueryException, IOException, OAuth2ClientAlreadyExistsForAppException,
+            io.supertokens.test.httpRequest.HttpResponseException, InterruptedException {
+
+        String[] args = {"../"};
+
+        this.process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        String clientId = "6030f07e-c8ef-4289-80c9-c18e0bf4f679-Not_Existing";
+        String propToChangeKey = "clientName";
+        String newValue = "Jozef2";
+
+        OAuthSQLStorage oAuthStorage = (OAuthSQLStorage) StorageLayer.getStorage(process.getProcess());
+
+        AppIdentifier testApp = new AppIdentifier("", "");
+        oAuthStorage.addClientForApp(testApp, clientId); // exists at our end, not exists in hydra
+
+        JsonObject requestBody = new JsonObject();
+        requestBody.addProperty("clientId", clientId);
+        requestBody.addProperty(propToChangeKey, newValue);
+
+        JsonObject response = HttpRequestForTesting.sendJsonPATCHRequest(process.getProcess(),
+                "http://localhost:3567/recipe/oauth/clients", requestBody);
+
+        assertEquals("OAUTH2_CLIENT_NOT_FOUND_ERROR", response.get("status").getAsString());
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+
 }